requirements.txt
discord.py
python-dotenv
dotdict
requests
aiohttp
loguru
tqdm